sidecar: Fix ping/pong response behavior in websocket client
Essentially we weren't responding to pings/pongs at all, as this is a separate
control setup within the websocket library. This meant that we were seeing a
timeout/reconnect every ~90s when we hit our read deadline.

The new logic appropriately sends a ping every 10s and we reset the deadline
when we receive back the corresponding pong.